% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_glance.R
\name{coeftable}
\alias{coeftable}
\alias{coeftable.lwdid_result}
\title{Extract coefficient table}
\usage{
coeftable(x, ...)

\method{coeftable}{lwdid_result}(x, ...)
}
\arguments{
\item{x}{A model object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Matrix or data.frame with coefficient information.
}
\description{
Generic function following fixest package conventions for extracting
a full coefficient table from model objects. For lwdid_result objects,
returns a matrix with estimates, standard errors, t-values, and p-values.
}
\seealso{
\code{\link{se}}, \code{\link{tstat}}, \code{\link{pvalue}},
\code{\link{tidy.lwdid_result}}

Other lwdid-fixest: 
\code{\link{pvalue}()},
\code{\link{se}()},
\code{\link{tstat}()}
}
\concept{lwdid-fixest}
