% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.build_group_columns}
\alias{.build_group_columns}
\title{Build groupby columns based on frequency}
\usage{
.build_group_columns(unit_var, time_var, frequency)
}
\description{
Build groupby columns based on frequency
}
\keyword{internal}
