% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{.can_use_parallel}
\alias{.can_use_parallel}
\title{Check whether parallel execution is available and active}
\usage{
.can_use_parallel()
}
\value{
Logical scalar.
}
\description{
Check whether parallel execution is available and active
}
\keyword{internal}
