% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.compute_cell_weighted_mean}
\alias{.compute_cell_weighted_mean}
\title{Compute weighted mean and optionally variance for a single cell}
\usage{
.compute_cell_weighted_mean(
  outcomes,
  raw_weights = NULL,
  compute_variance = FALSE
)
}
\description{
Compute weighted mean and optionally variance for a single cell
}
\keyword{internal}
