% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.compute_control_weighted_mean}
\alias{.compute_control_weighted_mean}
\title{Compute weighted mean for a control variable}
\usage{
.compute_control_weighted_mean(values, raw_weights = NULL)
}
\description{
Compute weighted mean for a control variable
}
\keyword{internal}
