% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.compute_effective_sample_size}
\alias{.compute_effective_sample_size}
\title{Compute effective sample size}
\usage{
.compute_effective_sample_size(weights)
}
\description{
Compute effective sample size
}
\keyword{internal}
