% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.count_periods}
\alias{.count_periods}
\title{Count unique periods in panel data}
\usage{
.count_periods(panel_data, time_var, frequency)
}
\description{
Count unique periods in panel data
}
\keyword{internal}
