% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_diagnostics.R
\name{.create_placebo_dataset}
\alias{.create_placebo_dataset}
\title{Create a placebo two-by-two dataset}
\usage{
.create_placebo_dataset(data, y, ivar, tvar, gvar, cohort, placebo_period)
}
\arguments{
\item{data}{data.frame or data.table.}

\item{y}{character(1). Outcome variable.}

\item{ivar}{character(1). Unit identifier.}

\item{tvar}{character(1). Time variable.}

\item{gvar}{character(1). Cohort variable.}

\item{cohort}{Numeric cohort identifier.}

\item{placebo_period}{Numeric placebo period.}
}
\value{
data.table containing placebo indicators.
}
\description{
Create a placebo two-by-two dataset
}
\keyword{internal}
