% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_diagnostics.R
\name{.determine_cluster_level}
\alias{.determine_cluster_level}
\title{Determine clustering level relative to unit}
\usage{
.determine_cluster_level(data, ivar, cluster_var)
}
\arguments{
\item{data}{data.frame or data.table.}

\item{ivar}{character(1). Unit identifier.}

\item{cluster_var}{character(1). Candidate clustering variable.}
}
\value{
character(1). One of "higher", "same", or "lower".
}
\description{
Determine clustering level relative to unit
}
\keyword{internal}
