% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{.determine_robustness_level}
\alias{.determine_robustness_level}
\title{Determine robustness level from sensitivity ratio}
\usage{
.determine_robustness_level(sensitivity_ratio)
}
\arguments{
\item{sensitivity_ratio}{numeric(1).}
}
\value{
character(1). Robustness level string.
}
\description{
Determine robustness level from sensitivity ratio
}
\keyword{internal}
