% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{.filter_sensitivity_args}
\alias{.filter_sensitivity_args}
\title{Filter extra arguments for a specific function}
\usage{
.filter_sensitivity_args(extra_args, fun)
}
\description{
Filter extra arguments for a specific function
}
\keyword{internal}
