% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.identify_mode}
\alias{.identify_mode}
\title{Identify estimation mode: Common Timing or Staggered}
\usage{
.identify_mode(d, post, gvar)
}
\value{
character(1). "common_timing" or "staggered".
}
\description{
Identify estimation mode: Common Timing or Staggered
}
\keyword{internal}
