% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.kahan_sum}
\alias{.kahan_sum}
\title{Kahan compensated summation (internal)}
\usage{
.kahan_sum(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
compensated sum
}
\description{
Kahan compensated summation (internal)
}
\keyword{internal}
