% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{.make_degenerate_period_row}
\alias{.make_degenerate_period_row}
\title{Construct degenerate period row (15 columns)}
\usage{
.make_degenerate_period_row(r, est)
}
\arguments{
\item{r}{Numeric, period value.}

\item{est}{List, return value from estimate_ra_common().}
}
\value{
data.frame with 1 row and 15 columns.
}
\description{
Construct degenerate period row (15 columns)
}
\keyword{internal}
