% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{.make_na_period_row}
\alias{.make_na_period_row}
\title{Construct all-NA period row (15 columns)}
\usage{
.make_na_period_row(r)
}
\arguments{
\item{r}{Numeric, period value.}
}
\value{
data.frame with 1 row and 15 columns.
}
\description{
Construct all-NA period row (15 columns)
}
\keyword{internal}
