% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{.summarize_transformation_comparison}
\alias{.summarize_transformation_comparison}
\title{Summarize transformation comparison arithmetic}
\usage{
.summarize_transformation_comparison(demean_estimate, detrend_estimate)
}
\description{
Summarize transformation comparison arithmetic
}
\keyword{internal}
