% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_choice}
\alias{.validate_choice}
\title{Validate that a parameter value is in a set of valid choices}
\usage{
.validate_choice(value, valid_set, name)
}
\description{
Validate that a parameter value is in a set of valid choices
}
\keyword{internal}
