% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_common_timing_structure}
\alias{.validate_common_timing_structure}
\title{Validate Common Timing specific panel structure}
\usage{
.validate_common_timing_structure(dt, ivar)
}
\arguments{
\item{dt}{data.table with tindex, d_, post_ columns}

\item{ivar}{character(1)}
}
\value{
list with K, tpost1, n_treated, n_control
}
\description{
Checks: post cross-unit consistency, post monotonicity,
pre/post period existence, treated/control existence.
}
\keyword{internal}
