% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_controls_columns}
\alias{.validate_controls_columns}
\title{Validate controls/ps_controls column existence}
\usage{
.validate_controls_columns(controls, data)
}
\description{
Validate controls/ps_controls column existence
}
\keyword{internal}
