% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_controls_dtype}
\alias{.validate_controls_dtype}
\title{Validate all control variables are numeric}
\usage{
.validate_controls_dtype(data, controls)
}
\arguments{
\item{data}{data.frame}

\item{controls}{character vector or NULL}
}
\value{
NULL (invisible).
}
\description{
Validate all control variables are numeric
}
\keyword{internal}
