% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_graph_options}
\alias{.validate_graph_options}
\title{Validate graph_options (NULL or named list)}
\usage{
.validate_graph_options(graph_options)
}
\description{
Validate graph_options (NULL or named list)
}
\keyword{internal}
