% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.validate_gvar_consistency}
\alias{.validate_gvar_consistency}
\title{Validate gvar consistency within units}
\usage{
.validate_gvar_consistency(data, unit_var, gvar)
}
\description{
Validate gvar consistency within units
}
\keyword{internal}
