% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_no_duplicates}
\alias{.validate_no_duplicates}
\title{Check for duplicate (ivar, tvar) observations}
\usage{
.validate_no_duplicates(dt, ivar, tvar)
}
\arguments{
\item{dt}{data.table}

\item{ivar}{character(1)}

\item{tvar}{character(1) or character(2)}
}
\value{
NULL (invisible). Raises lwdid_invalid_parameter if duplicates.
}
\description{
Check for duplicate (ivar, tvar) observations
}
\keyword{internal}
