% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_no_infinite_numeric_values}
\alias{.validate_no_infinite_numeric_values}
\title{Validate numeric variables do not contain infinite values}
\usage{
.validate_no_infinite_numeric_values(data, vars, estimator)
}
\arguments{
\item{data}{data.frame}

\item{vars}{character vector of variables to inspect.}

\item{estimator}{character scalar used in the error message.}
}
\value{
NULL (invisible).
}
\description{
Validate numeric variables do not contain infinite values
}
\keyword{internal}
