% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_nonneg_integer}
\alias{.validate_nonneg_integer}
\title{Validate non-negative integer (>= 0)}
\usage{
.validate_nonneg_integer(value, name)
}
\description{
Validate non-negative integer (>= 0)
}
\keyword{internal}
