% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_outcome_dtype}
\alias{.validate_outcome_dtype}
\title{Validate outcome variable is numeric (warn if logical)}
\usage{
.validate_outcome_dtype(data, y)
}
\arguments{
\item{data}{data.frame}

\item{y}{character(1). Outcome variable name.}
}
\value{
NULL (invisible).
}
\description{
Validate outcome variable is numeric (warn if logical)
}
\keyword{internal}
