% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_panel_balance}
\alias{.validate_panel_balance}
\title{Check panel balance and handle according to balanced_panel parameter}
\usage{
.validate_panel_balance(dt, ivar, balanced_panel, mode, gvar, tvar)
}
\arguments{
\item{dt}{data.table}

\item{ivar}{character(1)}

\item{balanced_panel}{character(1). "warn", "error", or "ignore".}

\item{mode}{character(1). "common_timing" or "staggered".}

\item{gvar}{character(1) or NULL.}

\item{tvar}{character(1) or character(2). Time variable(s).}
}
\value{
logical. TRUE if balanced, FALSE if unbalanced.
}
\description{
Check panel balance and handle according to balanced_panel parameter
}
\keyword{internal}
