% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_rolling_parameter}
\alias{.validate_rolling_parameter}
\title{Validate rolling parameter (case-insensitive) and seasonal requirements}
\usage{
.validate_rolling_parameter(rolling, tvar, season_var)
}
\value{
character(1). Standardized rolling parameter in lowercase.
}
\description{
Validate rolling parameter (case-insensitive) and seasonal requirements
}
\keyword{internal}
