% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_season_var_values}
\alias{.validate_season_var_values}
\title{Validate season_var column values are integers in 1..Q range}
\usage{
.validate_season_var_values(data, season_var, Q)
}
\description{
Validate season_var column values are integers in 1..Q range
}
\keyword{internal}
