% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validate_tvar}
\alias{.validate_tvar}
\title{Validate tvar parameter (length 1 or 2 character vector)}
\usage{
.validate_tvar(tvar, data)
}
\description{
Validate tvar parameter (length 1 or 2 character vector)
}
\keyword{internal}
