% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.validate_weights}
\alias{.validate_weights}
\title{Validate weights: non-negative, handle missing}
\usage{
.validate_weights(data, weight_var)
}
\description{
Validate weights: non-negative, handle missing
}
\keyword{internal}
