% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_cross_section}
\alias{extract_cross_section}
\title{Extract cross-section for a single period}
\usage{
extract_cross_section(dt, tvar, period)
}
\arguments{
\item{dt}{data.table, panel data.}

\item{tvar}{Character, name of the time variable column.}

\item{period}{Scalar numeric, target period.}
}
\value{
data.table subset with rows where \code{tvar == period}.
}
\description{
Extract cross-section for a single period
}
\keyword{internal}
