% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cohorts}
\alias{get_cohorts}
\title{Extract sorted unique treatment cohorts from panel data}
\usage{
get_cohorts(data, gvar, ivar, never_treated_values = c(0, Inf))
}
\arguments{
\item{data}{data.frame or data.table, panel data.}

\item{gvar}{Character, name of the cohort variable column.}

\item{ivar}{Character, name of the unit identifier column.}

\item{never_treated_values}{Numeric vector of explicit NT marker values.}
}
\value{
Sorted integer vector of valid cohort values.
}
\description{
Extract sorted unique treatment cohorts from panel data
}
\keyword{internal}
