% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_glance.R
\name{glance}
\alias{glance}
\title{Re-export glance generic from generics}
\usage{
glance(x, ...)
}
\arguments{
\item{x}{An object to glance at}

\item{...}{Additional arguments passed to methods}
}
\description{
Re-export glance generic from generics
}
