% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_never_treated}
\alias{has_never_treated}
\title{Check if panel data contains never-treated units}
\usage{
has_never_treated(data, gvar, ivar)
}
\arguments{
\item{data}{data.frame or data.table, panel data.}

\item{gvar}{Character, name of the cohort variable column.}

\item{ivar}{Character, name of the unit identifier column.}
}
\value{
Single logical value.
}
\description{
Check if panel data contains never-treated units
}
\keyword{internal}
