% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{new_lwdid_aggregation_result}
\alias{new_lwdid_aggregation_result}
\title{Create an lwdid_aggregation_result S3 object}
\usage{
new_lwdid_aggregation_result(
  panel_data,
  n_original_obs,
  n_cells,
  n_units,
  n_periods,
  cell_stats,
  min_cell_size_stat,
  max_cell_size,
  mean_cell_size,
  median_cell_size,
  unit_var,
  time_var,
  outcome_var,
  weight_var,
  frequency,
  n_excluded_cells = 0L,
  excluded_cells_info = list()
)
}
\arguments{
\item{panel_data}{data.table, aggregated panel data}

\item{n_original_obs}{integer, total observations in original data}

\item{n_cells}{integer, number of (unit, period) cells}

\item{n_units}{integer, number of unique units}

\item{n_periods}{integer, number of unique periods}

\item{cell_stats}{data.table, cell-level statistics}

\item{min_cell_size_stat}{integer, minimum cell size}

\item{max_cell_size}{integer, maximum cell size}

\item{mean_cell_size}{numeric, mean cell size}

\item{median_cell_size}{numeric, median cell size}

\item{unit_var}{character, unit variable name}

\item{time_var}{character, time variable name(s)}

\item{outcome_var}{character, outcome variable name}

\item{weight_var}{character or NULL, weight variable name}

\item{frequency}{character, aggregation frequency}

\item{n_excluded_cells}{integer, number of excluded cells}

\item{excluded_cells_info}{list, excluded cell details}
}
\value{
lwdid_aggregation_result S3 object
}
\description{
Create an lwdid_aggregation_result S3 object
}
