% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{nobs.lwdid_result}
\alias{nobs.lwdid_result}
\title{Number of observations}
\usage{
\method{nobs}{lwdid_result}(object, ...)
}
\arguments{
\item{object}{lwdid_result object}

\item{...}{ignored}
}
\value{
integer, number of observations
}
\description{
Extract the number of observations used in the lwdid model.
Ensures compatibility with modelsummary and broom::glance().
}
