% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.lwdid_clustering_diagnosis}
\alias{plot.lwdid_clustering_diagnosis}
\title{Plot clustering diagnostic}
\usage{
\method{plot}{lwdid_clustering_diagnosis}(x, ...)
}
\arguments{
\item{x}{lwdid_clustering_diagnosis object}

\item{...}{passed to ggplot2}
}
\value{
ggplot object
}
\description{
Plot cluster size distribution for a
\code{lwdid_clustering_diagnosis} object.
}
