% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.lwdid_selection_diagnosis}
\alias{plot.lwdid_selection_diagnosis}
\title{Plot selection diagnostic}
\usage{
\method{plot}{lwdid_selection_diagnosis}(x, ...)
}
\arguments{
\item{x}{lwdid_selection_diagnosis object}

\item{...}{passed to ggplot2}
}
\value{
ggplot object
}
\description{
Plot attrition over time for a
\code{lwdid_selection_diagnosis} object.
}
