% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot_event_study}
\alias{plot_event_study}
\title{Event Study Plot}
\usage{
plot_event_study(x, ...)
}
\arguments{
\item{x}{An object to plot.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A ggplot object or data.
}
\description{
Generic function for event study visualization.
}
