% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_diagnostics.R
\name{print.lwdid_parallel_trends}
\alias{print.lwdid_parallel_trends}
\title{Print method for parallel trends diagnostics}
\usage{
\method{print}{lwdid_parallel_trends}(x, ...)
}
\arguments{
\item{x}{An object of class \code{lwdid_parallel_trends}.}

\item{...}{Additional arguments (ignored).}
}
\value{
\code{x} invisibly.
}
\description{
Print method for parallel trends diagnostics
}
