% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{print.lwdid_sensitivity}
\alias{print.lwdid_sensitivity}
\title{Print method for lwdid_sensitivity}
\usage{
\method{print}{lwdid_sensitivity}(x, ...)
}
\arguments{
\item{x}{An lwdid_sensitivity object.}

\item{...}{Additional arguments (ignored).}
}
\value{
x invisibly.
}
\description{
Print method for lwdid_sensitivity
}
