% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{print.lwdid_sensitivity_comprehensive}
\alias{print.lwdid_sensitivity_comprehensive}
\title{Print method for lwdid_sensitivity_comprehensive}
\usage{
\method{print}{lwdid_sensitivity_comprehensive}(x, ...)
}
\arguments{
\item{x}{An lwdid_sensitivity_comprehensive object.}

\item{...}{Additional arguments (ignored).}
}
\value{
x invisibly.
}
\description{
Print method for lwdid_sensitivity_comprehensive
}
