% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wild_bootstrap.R
\name{print.lwdid_wcb_result}
\alias{print.lwdid_wcb_result}
\title{Print a wild cluster bootstrap result}
\usage{
\method{print}{lwdid_wcb_result}(x, digits = 4L, ...)
}
\arguments{
\item{x}{An object of class \code{lwdid_wcb_result}.}

\item{digits}{Number of printed digits.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a wild cluster bootstrap result
}
