% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_glance.R
\name{pvalue}
\alias{pvalue}
\alias{pvalue.lwdid_result}
\title{Extract p-values}
\usage{
pvalue(x, ...)

\method{pvalue}{lwdid_result}(x, ...)
}
\arguments{
\item{x}{A model object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Named numeric vector of p-values.
}
\description{
Generic function following fixest package conventions for extracting
p-values from model objects. For lwdid_result objects, returns
the p-values of the ATT estimate(s).
}
\seealso{
\code{\link{se}}, \code{\link{tstat}}, \code{\link{coeftable}},
\code{\link{tidy.lwdid_result}}

Other lwdid-fixest: 
\code{\link{coeftable}()},
\code{\link{se}()},
\code{\link{tstat}()}
}
\concept{lwdid-fixest}
