% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{summary.lwdid_sensitivity_comprehensive}
\alias{summary.lwdid_sensitivity_comprehensive}
\title{Summary method for lwdid_sensitivity_comprehensive}
\usage{
\method{summary}{lwdid_sensitivity_comprehensive}(object, ...)
}
\arguments{
\item{object}{An lwdid_sensitivity_comprehensive object.}

\item{...}{Additional arguments (ignored).}
}
\value{
object invisibly.
}
\description{
Summary method for lwdid_sensitivity_comprehensive
}
