% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_glance.R
\name{tidy}
\alias{tidy}
\title{Re-export tidy generic from generics}
\usage{
tidy(x, ...)
}
\arguments{
\item{x}{An object to tidy}

\item{...}{Additional arguments passed to methods}
}
\description{
Re-export tidy generic from generics
}
