% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{to_dict}
\alias{to_dict}
\title{Convert an object to a dictionary (named list)}
\usage{
to_dict(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A named list of parameters and statistics.
}
\description{
Convert an object to a dictionary (named list)
}
