% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{to_latex}
\alias{to_latex}
\title{Export an object to LaTeX table format}
\usage{
to_latex(x, ...)
}
\arguments{
\item{x}{An object to export.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A character string containing LaTeX table markup.
}
\description{
S3 generic for converting estimation results to LaTeX table strings.
}
