% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_glance.R
\name{tstat}
\alias{tstat}
\alias{tstat.lwdid_result}
\title{Extract t-statistics}
\usage{
tstat(x, ...)

\method{tstat}{lwdid_result}(x, ...)
}
\arguments{
\item{x}{A model object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Named numeric vector of t-statistics.
}
\description{
Generic function following fixest package conventions for extracting
t-statistics from model objects. For lwdid_result objects, returns
the t-statistics of the ATT estimate(s).
}
\seealso{
\code{\link{se}}, \code{\link{pvalue}}, \code{\link{coeftable}},
\code{\link{tidy.lwdid_result}}

Other lwdid-fixest: 
\code{\link{coeftable}()},
\code{\link{pvalue}()},
\code{\link{se}()}
}
\concept{lwdid-fixest}
