% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{vcov.lwdid_result}
\alias{vcov.lwdid_result}
\title{Variance-covariance matrix}
\usage{
\method{vcov}{lwdid_result}(object, type = "overall", ...)
}
\arguments{
\item{object}{lwdid_result object}

\item{type}{character, granularity: "overall" (1x1), "by_period",
"by_cohort", "full" (complete OLS parameter VCE)}

\item{...}{ignored}
}
\value{
variance-covariance matrix
}
\description{
Extract variance-covariance matrix for ATT estimates.
}
